Crate quorum_core

This file is the main entry point for the quorum-core library. The quorum-core is includes all code which is shared between both public and private servers.

For example this includes the health and echo route endpoints since both public and private servers have them exactly the same,

Modules

cli

Main entry point for the server cli Commands This is where commands are defined properly, and where command functions are called.

db

models

routes

startup

Server startup printing initiliation Provides public functions to be used for neat and pretty output during server startup. Including the ASCII banner, step-by-step progress logging, and timing information.

utility

Macros

define_config_schema