quorum_core::utility

Module std

This file includes some basic std::io functionality including outputting text as a typewriter effect and waiting for user input to continue.

Functions

press_enter_to_continue

Waits for the user to press Enter to continue. Optionally clears the screen and reprints the banner.

typewriter_println

Prints the given text to the console with a typewriter effect, where each character is printed with a delay.